FEMA tools update.
Working on filtering out relevant data for charities.

Currently, filters by beginning year (ugly implementation rn).

Seems to be a maximum page size of 1000 disasters, and filtering just for 2021 still reaches this limit. Will eventually need to check for severity to only consider major disasters. Either that or find a way to circumvent the max page size.

Added an abstract method for ApiQueries to handle the response generated by the HTML get request.