Added query filtering for disasters.
Subclasses of ApiQuery now have to handler responses to themselves.

DisasterQuery now has fields and declaration types defined for itself. The declaration types can be used in its filter and the fields can be used to access the data returned from the api. This eliminates the need for a data class for each query. It can simply be accessed with json indexing using those fields.

DisasterQuery can now take python datetime to filter the query to after a specific date.

The disaster query class now generates a filter for any given parameters.